begun to separate functions to MapperValued
